app/requirements-dev.txt
black==20.8b1
ipdb
ipython
prospector==1.3.1
pydevd
pyramid_ipython
pytest==6.2.1
webtest==2.0.35
responses
